+++ src/Exception/Exception.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace GraphQL\Exception;
+
+/**
+ * Interface Exception
+ *
+ * @package GraphQL\Exception
+ */
+interface Exception
+{
+}