+++ Changelog.md
@@ -45,6 +45,15 @@ constructing a new query and adding a nested subfield
 - Updated Query class to allow for an alias
 - Updated QueryBuilder class to allow for an alias
 
+## 2.0
+
+### Changed
+
+- Changed hard Guzzle dependency any PSR-18 implementation.
+- Added dependency on a PSR-7 implementation. If you are using Guzzle it requires guzzlehttp/psr7 which will fill this dependency.
+- Exceptions returned from Client are now in the `GraphQL\Exception\Client` namespace instead of Guzzle.
+- The only httpOptions that are accepted are `headers` and `version`.
+
 ## 1.7:
 
 ### Added